Fix enum tests for PHP 7.4/8.0 compatibility
