+++ tests/framework/helpers/enums/IntBackedEnum.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace yiiunit\framework\helpers\enums;
+
+enum IntBackedEnum: int
+{
+    case Active = 1;
+    case Inactive = 0;
+}
