+++ tests/framework/helpers/enums/PureEnum.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace yiiunit\framework\helpers\enums;
+
+enum PureEnum
+{
+    case Red;
+    case Blue;
+}
