+++ tests/framework/helpers/enums/StringBackedEnum.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace yiiunit\framework\helpers\enums;
+
+enum StringBackedEnum: string
+{
+    case Hearts = 'hearts';
+    case Diamonds = 'diamonds';
+}
